volumes/mounts: remove LCOW code
Both "windows" and "linux" parsers were previously compiled on all platforms,
so that both semantics could be handled; now that LCOW is removed, renaming these
files back to be compiled only on the matching platform.

With the OS-specific handling being removed, the "OS" parameter is also removed
from `daemon.verifyContainerSettings()`.
